Changed the collector so that it can have multiple independent collections ISSUE=5706
As part of this, instead of having indpendent <collection> and <hosts> tags,
the functions of <hosts> have been merged into <collection> to allow different
collections to have independent sets of hosts.

In addition, I moved the location of _set_hup to make _load_config cleaner,
and changed the worker-name scheme to reduce a potential source of ambiguity.